pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aioboto3>=13.0.0",
    "pypdf>=4.0.0",
    "tiktoken>=0.7.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "prometheus-client>=0.20.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "faker>=25.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
